Refactor
Refactor test case
