+++ src/main/java/com/group/artifact/web/HomeController.java
@@ -1,9 +1,7 @@
 package com.group.artifact.web;
 
-import com.group.artifact.vo.Token;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
